Refactor EditableField
